[feat] #35 Shared 프로퍼티 래퍼 구현
